+++ libft/ft_ultoa_base.c
@@ -6,7 +6,7 @@
 /*   By: nfukuma <nfukuma@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/16 23:48:06 by nfukuma           #+#    #+#             */
-/*   Updated: 2022/07/19 00:01:59 by nfukuma          ###   ########.fr       */
+/*   Updated: 2022/07/19 17:05:46 by nfukuma          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 static char	*ft_recursive_create_str(char *str, unsigned long n, int base,
 		char *base_set)
 {
-	if (n >= base)
+	if (n >= (unsigned long)base)
 		str = ft_recursive_create_str(str, (n / base), base, base_set);
 	*str++ = base_set[n % base];
 	return (str);
